+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.20.1-4.1.0.1] - 2023-11-04
+### Changed
+- Updated compat with Moonlight Lib to 1.20-2.8.54
+
+### Fixed
+- Every Compat/Moonlight Lib wood blocks were not available (thanks to Zartrix for the report & MehVahdJukaar for the support fixing the issue) #81 
+
 ## [1.20.1-4.1.0.0] - 2023-08-09
 ### Changed
 - Changed Forge to NeoForge 1.20.1-47.1.54 (compatible with Forge 47.1.0)